Fix escape newline in note from old version

+++ packages/lasotuvi_domain/lib/src/storage/remover.dart
@@ -1,5 +1,6 @@
 import 'package:lasotuvi_domain/src/chart/usecase/delete_avatar_from_cloud.dart';
 import 'package:tauari_data_core/tauari_data_core.dart';
+import 'package:tauari_utils/tauari_utils.dart';
 
 import '../chart/entity/chart.dart';
 import '../chart/usecase/delete_chart_from_cloud.dart';
@@ -22,6 +23,7 @@ class Remover {
   final RefreshCache refreshChartCloud;
   final RefreshCache refreshTagCloud;
   final RefreshCache refreshNoteCloud;
+  final RefreshCache refreshChartTagCloud;
 
   const Remover({
     required this.deleteChartFromLocal,
@@ -34,6 +36,7 @@ class Remover {
     required this.refreshChartCloud,
     required this.refreshTagCloud,
     required this.refreshNoteCloud,
+    required this.refreshChartTagCloud,
   });
   Future<void> deleteFromLocal<T>({required Iterable<T> items}) async {
     if (T == Chart) {
@@ -62,11 +65,16 @@ class Remover {
           item,
           false,
         );
-        if (!(item.avatar == null || item.avatar!.isEmpty)) {
-          await deleteAvatarFromCloud(uid, item.avatar!.split('/').last);
+        if (await availableNetwork()) {
+          try {
+            if (!(item.avatar == null || item.avatar!.isEmpty)) {
+              await deleteAvatarFromCloud(uid, item.avatar!);
+            }
+          } catch (_) {}
         }
       }
       refreshChartCloud();
+      refreshChartTagCloud();
     } else if (T == Tag) {
       for (var i = 0; i < items.length; i++) {
         await deleteTagFromCloud(
@@ -76,6 +84,7 @@ class Remover {
         );
       }
       refreshTagCloud();
+      refreshChartTagCloud();
     } else if (T == Note) {
       for (var i = 0; i < items.length; i++) {
         await deleteNoteFromCloud(uid, items.elementAt(i) as Note, false);
@@ -85,35 +94,41 @@ class Remover {
   }
 
   Future<void> deleteForever<T>(
-      {required String uid, required Iterable<T> items}) async {
+      {required String? uid, required Iterable<T> items}) async {
     if (T == Chart) {
       for (var i = 0; i < items.length; i++) {
         await deleteChartFromLocal(items.elementAt(i) as Chart);
-        await deleteChartFromCloud(
-          uid,
-          items.elementAt(i) as Chart,
-          false,
-        );
+        if (uid != null) {
+          await deleteChartFromCloud(
+            uid,
+            items.elementAt(i) as Chart,
+            false,
+          );
+        }
       }
       refreshChartCloud();
     } else if (T == Tag) {
       for (var i = 0; i < items.length; i++) {
         await deleteTagFromLocal(items.elementAt(i) as Tag);
-        await deleteTagFromCloud(
-          uid,
-          items.elementAt(i) as Tag,
-          false,
-        );
+        if (uid != null) {
+          await deleteTagFromCloud(
+            uid,
+            items.elementAt(i) as Tag,
+            false,
+          );
+        }
       }
       refreshTagCloud();
     } else if (T == Note) {
       for (var i = 0; i < items.length; i++) {
         await deleteNoteFromLocal(items.elementAt(i) as Note);
-        await deleteNoteFromCloud(
-          uid,
-          items.elementAt(i) as Note,
-          false,
-        );
+        if (uid != null) {
+          await deleteNoteFromCloud(
+            uid,
+            items.elementAt(i) as Note,
+            false,
+          );
+        }
       }
       refreshNoteCloud();
     }
